requirements.txt
streamlit>=1.35.0
anthropic>=0.25.0
chromadb>=0.5.0
faster-whisper>=1.0.0
gTTS>=2.5.0
pdfplumber>=0.11.1
httpx>=0.27.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
python-dotenv>=1.0.0
